ENH: Added dwi_nrrd module to pynrrd
dwi_nrrd provides important elements to deal with DWI meta data.
This module provide support for DWI imaging data in pynrrd.
Additional tests are provided to show and verify the functionality
of this new module.
