+++ dwi_nrrd.py
@@ -1,22 +1,9 @@
 """
-testNrrd.py
-==============
-FIRST, DWIConvert nrrd to nii.gz
-SECOND, Use existing nibabel to read .nii.gz
-THIRD, Use new nrrd.py to read .nrrd file
-VERIFY that the pynrrd can be converted to nibabel compliant data model
-FINAL, do something interesting with manipulating DWI data (i.e. average close gradients)
-Last, write averaged nrrd file to disk
+dwi_nrrd.py
 
-Usage:
-  testNrrd.py --inputDWIScan DWISCAN --outputDir OUTPUTDIR --NEP_build_path PROGRAM_PATH
+A python implementation that provides important elements for using DWI imaging data in nrrd format.
+See http://teem.sourceforge.net/nrrd/format.html for the specification.
 
-Options:
-  -h --help                             Show this help and exit
-  -v --version                          Print the version and exit
-  --inputDWIScan DWISCAN                Path to the test DWI scan
-  --outputDir OUTPUTDIR                 Path to your test directory
-  --NEP_build_path PROGRAM_PATH         Path to last NEP build for DWIConvert
 """
 
 __author__ = 'johnsonhj,aghayoor'
@@ -157,134 +144,3 @@ def WriteNAMICDWIToNrrd(filename, data, bvecs, bvals, options=None):
     options['keyvaluepairs']=keyvaluePairDict
     options['encoding']='gzip' # Always use gzip compression for DWI data
     nrrd.write(filename,data,options)
-
-
-def unit_vector(vector):
-  """ Returns the unit vector of the vector.  """
-  v_norm = np.linalg.norm(vector)
-  if v_norm < 1e-4: ## allow for small tolerance
-    return vector
-  else:
-    return vector / v_norm
-
-def angle_between(v1, v2):
-  """ Returns the angle in radians between vectors 'v1' and 'v2' """
-  v1_u = unit_vector(v1)
-  v2_u = unit_vector(v2)
-  angle = np.arccos(np.dot(v1_u, v2_u))
-  if np.isnan(angle):
-    if (v1_u == v2_u).all():
-      return 0.0
-    else:
-      return np.pi
-  return angle
-
-def angle_degrees(v1,v2):
-  ang_rad=angle_between(v1,v2)
-  return ang_rad*180/np.pi
-
-def average(v1,v2):
-  return np.mean( np.array([ v1, v2 ]), axis=0 )
-
-def AverageLikeGradients(nrrd_data,nrrd_bvecs,nrrd_bvals,gradient_index):
-    starting_num_bvecs = len(nrrd_bvecs)
-    processed_list = [ False ] * starting_num_bvecs
-    remove_list = [ False ] * starting_num_bvecs
-    for i in range(0,starting_num_bvecs):
-       processed_list[i] = True
-       for j in range(i,starting_num_bvecs):
-         if processed_list[j] == False:
-           if( angle_degrees(nrrd_bvecs[i], nrrd_bvecs[j] ) < 3
-               and abs(nrrd_bvals[i] - nrrd_bvals[j] ) < 1 ):
-                  nrrd_bvecs[i] = average(nrrd_bvecs[i], nrrd_bvecs[j] )
-                  nrrd_bvals[i] = average(nrrd_bvals[i], nrrd_bvals[j] )
-                  processed_list[j] = True
-                  remove_list[j] = True
-
-
-    # http://docs.scipy.org/doc/numpy/reference/generated/numpy.delete.html
-    remove_indices = []
-    for index in range(0,len(remove_list)):
-      if remove_list[index]:
-         remove_indices.append(index)
-
-    nrrd_bvecs = np.delete( nrrd_bvecs, remove_indices, 0 )
-    nrrd_bvals = np.delete( nrrd_bvals, remove_indices, 0 )
-    if(gradient_index == -1): gradient_index = 0
-    nrrd_data = np.delete( nrrd_data, remove_indices, gradient_index)
-    return nrrd_data, nrrd_bvecs, nrrd_bvals
-
-#######################
-#######################
-#######################
-#######################
-#######################
-#######################
-# Testing Code Here
-if __name__ == '__main__':
-  import os
-  import sys
-  import subprocess
-  import nibabel as nib
-  from dipy.io import read_bvals_bvecs
-
-  from docopt import docopt
-  argv = docopt(__doc__, version='1.0')
-  print argv
-
-  DWISCAN = argv['--inputDWIScan']
-  assert os.path.exists(DWISCAN), "Input DWI scan is not found: %s" % DWISCAN
-
-  PROGRAM_PATH = argv['--NEP_build_path']
-
-  if argv['--outputDir'] == None:
-    print("*** Output directory is set to current working directory.")
-    OUTPUTDIR = os.getcwd()
-  else:
-    OUTPUTDIR = argv['--outputDir']
-    assert os.path.exists(OUTPUTDIR), "Output directory is not found: %s" % OUTPUTDIR
-
-  # FIRST use nibabel to read nifti formatted DWI data
-  print("Making some nifti formatted data")
-  arg1='--inputVolume'+' '+str(DWISCAN)
-  arg2='--conversionMode NrrdToFSL'
-  arg3='--outputVolume'+' '+os.path.join(OUTPUTDIR,'testFSL.nii.gz')
-  subprocess.call([os.path.join(PROGRAM_PATH,'DWIConvert'), arg1, arg2, arg3])
-
-  # SECOND
-  nifti_file=os.path.join(OUTPUTDIR,'testFSL.nii.gz')
-  fbvec=os.path.join(OUTPUTDIR,'testFSL.bvec')
-  fbval=os.path.join(OUTPUTDIR,'testFSL.bval')
-
-  img = nib.load(nifti_file)
-  data = img.get_data()
-
-  nifti_bvals, nifti_bvecs = read_bvals_bvecs(fbval, fbvec)
-  print nifti_bvals
-  print nifti_bvecs
-  print type(nifti_bvals)
-  print type(nifti_bvecs)
-
-  #THIRD use nrrd.py + new header parser to read data
-  nrrd_data,myOptions,nrrd_bvecs,nrrd_bvals, gradient_index = ReadNAMICDWIFromNrrd(DWISCAN)
-
-  print type(nrrd_bvecs)
-  print type(nrrd_bvals)
-
-  print np.subtract(nifti_bvecs,nrrd_bvecs)
-  print np.subtract(nifti_bvals,nrrd_bvals)
-
-  # FINAL
-  # Write an algorithm purely in python to average similar gradients together
-  #
-  print("Number of components BEFORE averaging: {0}".format(len(nrrd_bvecs)))
-
-  nrrd_data,nrrd_bvecs,nrrd_bvals=AverageLikeGradients(nrrd_data,nrrd_bvecs,nrrd_bvals,gradient_index)
-
-  print("Number of components AFTER averaging: {0}".format(len(nrrd_bvecs)))
-  #print("4th dimension of nrrd_data after averaging: {0}".format(len(nrrd_data[:,:,:,1])))
-  outputNrrdFilename=os.path.join(OUTPUTDIR,'averagedNrrdFile.nrrd')
-  WriteNAMICDWIToNrrd(outputNrrdFilename,nrrd_data,nrrd_bvecs,nrrd_bvals,myOptions)
-  print("Output nrrd file is written to the disk: {0}.".format(outputNrrdFilename))
-
-  sys.exit(0)
