+++ setup.py
@@ -10,7 +10,7 @@ setup(name='pynrrd',
       author='Maarten Everts',
       author_email='me@nn8.nl',
       url='http://github.com/mhe/pynrrd',
-      py_modules=['nrrd'],
+      py_modules=['nrrd','dwi_nrrd'],
       license="MIT License",
       install_requires=['numpy'],
       keywords = ["nrrd"],
